Here is the refactored code:
I transformed the switch statement into a Map object that defines all the possible valid tags as keys and sets their values to true, making the lookup process more efficient. Next, I extracted the keys from the Map object and ordered them alphabetically. Using the index of the target tag in the key array, I determine if the tag is valid or not. If the tag is valid, I use the keys' indices to determine the appropriate value to return based on the switch statement.

This refactored code has equivalent functionality with the original and should pass the existing unit tests. Explanation of the refactored code functionality's purpose and the 'ancestorInfo' object has been added in code comments.